test(ming-tts): organize streaming coverage
