+++ tests/unit_test/ming_tts/test_engine_builder.py
@@ -1,0 +1,19 @@
+# SPDX-License-Identifier: Apache-2.0
+
+from __future__ import annotations
+
+from sglang_omni.models.ming_tts.engine_builder import MingTtsEngineBuilder
+from sglang_omni.models.ming_tts.model_runner import MingTTSModelRunner
+
+
+def test_ming_tts_abort_callback_resets_runner_state() -> None:
+    runner = object.__new__(MingTTSModelRunner)
+    runner._request_states = {"req-ming-tts": object()}
+    builder = object.__new__(MingTtsEngineBuilder)
+    builder._model_runner = runner
+
+    abort_callback = builder.make_abort_callback()
+    abort_callback("req-ming-tts")
+    abort_callback("req-ming-tts")
+
+    assert runner._request_states == {}
